Initial testing of ContractImpl.findBooking(BookingIdentifier)

+++ src/main/java/backend/datalayer/dao/CustomBookingRepository.java
@@ -7,7 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.NoRepositoryBean;
 
 public interface CustomBookingRepository {
-    // Skal måske wrappes i noget andet, så det ikke er DB objekt der returneres...
-    BookingDB findBooking(int id);
+    BookingDB findBooking(Long id);
 
 }
